+++ src/test/java/us/ihmc/euclid/tuple3D/Vector3DTest.java
@@ -81,7 +81,7 @@ public class Vector3DTest extends Vector3DBasicsTest<Vector3D>
       {
          tuple1.setElement(i % 3, random.nextDouble());
          newHashCode = tuple1.hashCode();
-         assertNotEquals((long) newHashCode, (long) previousHashCode);
+         assertNotEquals(newHashCode, previousHashCode);
          previousHashCode = newHashCode;
       }
    }
