+++ src/test/java/us/ihmc/euclid/tuple4D/QuaternionTest.java
@@ -135,7 +135,7 @@ public class QuaternionTest extends QuaternionBasicsTest<Quaternion>
          }
          q.setUnsafe(qx, qy, qz, qs);
          newHashCode = q.hashCode();
-         assertNotEquals((long) newHashCode, (long) previousHashCode);
+         assertNotEquals(newHashCode, previousHashCode);
          previousHashCode = newHashCode;
       }
    }
